start kubeapiserver package for sharing between kubeapiserver and federation
