+++ pkg/genericapiserver/authorizer/BUILD
@@ -17,11 +17,7 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//pkg/auth/authorizer:go_default_library",
-        "//pkg/auth/authorizer/abac:go_default_library",
-        "//pkg/auth/authorizer/union:go_default_library",
         "//pkg/client/clientset_generated/clientset/typed/authorization/v1beta1:go_default_library",
-        "//pkg/controller/informers:go_default_library",
-        "//plugin/pkg/auth/authorizer/rbac:go_default_library",
         "//plugin/pkg/auth/authorizer/webhook:go_default_library",
     ],
 )
